rename: 파일명 컴포넌트명과 동일하도록 수정
