+++ app/map/components/BottomSheetContent.tsx
