+++ app/map/components/BottomSheetHeader.tsx
